requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
django-environ==0.12.0
djangorestframework==3.16.0
idna==3.10
iniconfig==2.1.0
mysqlclient==2.2.7
oauthlib==3.2.2
packaging==25.0
pluggy==1.6.0
pytest==8.3.5
pytest-django==4.11.1
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.3
tweepy==4.15.0
urllib3==2.4.0
